requirements.txt
google-api-python-client~=2.172.0
google-auth-httplib2
google-auth-oauthlib~=1.2.2
oauthlib~=3.3.0
protobuf~=6.31.1
requests~=2.32.4
